chore(refactor): extract class
